mapauto-microservices/osscan_service/requirements.txt
# =Juggernaut= Updated requirements.txt with all necessary dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
# Standard library modules (included in Python by default):
# os, subprocess, shlex, ipaddress, typing, datetime, pathlib
